refactored navbar - header and footer
